requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pandas==2.1.4
tabula-py==2.9.0
pdfplumber==0.10.3
python-multipart==0.0.6
